[giac] attempt to fix pointer cast
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@45626 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ giac/src/giac/cpp/gen.cc
@@ -12541,7 +12541,7 @@ namespace giac {
       return print_EQW(*_EQWptr);
     case _POINTER_: {
       // handle 64 bits pointers
-      unsigned long u=(unsigned long)_POINTER_val;
+      unsigned long long u=(unsigned long long)_POINTER_val;
       if (u<(1U<<31))
 	return "pointer("+hexa_print_INT_(int((alias_type)_POINTER_val))+","+print_INT_(subtype)+")";
       gen z=longlong(u);
